Move BulletShapeManager and BulletShape to resource/
