+++ components/resource/bulletshapemanager.hpp
@@ -6,6 +6,8 @@
 
 #include <osg/ref_ptr>
 
+#include "bulletshape.hpp"
+
 namespace VFS
 {
     class Manager;
@@ -14,10 +16,6 @@ namespace VFS
 namespace Resource
 {
     class SceneManager;
-}
-
-namespace NifBullet
-{
 
     class BulletShape;
     class BulletShapeInstance;
